# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SIT ORBIT API",
    "description": "Personal Campus Agent for Shibaura Institute of Technology",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/agent/runs": {
      "post": {
        "summary": "Start Agent Run",
        "operationId": "start_agent_run_v1_agent_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/AgentRunCompleted"
                    },
                    {
                      "$ref": "#/components/schemas/AgentRunToolRequired"
                    }
                  ],
                  "title": "Response Start Agent Run V1 Agent Runs Post",
                  "discriminator": {
                    "propertyName": "status",
                    "mapping": {
                      "completed": "#/components/schemas/AgentRunCompleted",
                      "tool_required": "#/components/schemas/AgentRunToolRequired"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/agent/runs/{run_id}/tool-results": {
      "post": {
        "summary": "Submit Agent Tool Result",
        "operationId": "submit_agent_tool_result_v1_agent_runs__run_id__tool_results_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentToolResultRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/AgentRunCompleted"
                    },
                    {
                      "$ref": "#/components/schemas/AgentRunToolRequired"
                    }
                  ],
                  "discriminator": {
                    "propertyName": "status",
                    "mapping": {
                      "completed": "#/components/schemas/AgentRunCompleted",
                      "tool_required": "#/components/schemas/AgentRunToolRequired"
                    }
                  },
                  "title": "Response Submit Agent Tool Result V1 Agent Runs  Run Id  Tool Results Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/chat/runs": {
      "post": {
        "summary": "Start Chat Run",
        "operationId": "start_chat_run_v1_chat_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ChatRunCompleted"
                    },
                    {
                      "$ref": "#/components/schemas/ChatRunToolRequired"
                    }
                  ],
                  "title": "Response Start Chat Run V1 Chat Runs Post",
                  "discriminator": {
                    "propertyName": "status",
                    "mapping": {
                      "completed": "#/components/schemas/ChatRunCompleted",
                      "tool_required": "#/components/schemas/ChatRunToolRequired"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/chat/runs/{run_id}/tool-results": {
      "post": {
        "summary": "Submit Chat Tool Result",
        "operationId": "submit_chat_tool_result_v1_chat_runs__run_id__tool_results_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatToolResultRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ChatRunCompleted"
                    },
                    {
                      "$ref": "#/components/schemas/ChatRunToolRequired"
                    }
                  ],
                  "discriminator": {
                    "propertyName": "status",
                    "mapping": {
                      "completed": "#/components/schemas/ChatRunCompleted",
                      "tool_required": "#/components/schemas/ChatRunToolRequired"
                    }
                  },
                  "title": "Response Submit Chat Tool Result V1 Chat Runs  Run Id  Tool Results Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/actions/propose": {
      "post": {
        "summary": "Propose Action",
        "operationId": "propose_action_v1_actions_propose_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposeActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionProposal"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/actions/{action_id}/verify": {
      "post": {
        "summary": "Verify Action",
        "operationId": "verify_action_v1_actions__action_id__verify_post",
        "parameters": [
          {
            "name": "action_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Action Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrbitEvent"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ActionProposal": {
        "properties": {
          "action_id": {
            "type": "string",
            "minLength": 1,
            "title": "Action Id"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "title": "Title"
          },
          "reason": {
            "type": "string",
            "minLength": 1,
            "title": "Reason"
          },
          "duration_minutes": {
            "type": "integer",
            "maximum": 180.0,
            "minimum": 1.0,
            "title": "Duration Minutes"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceLink"
            },
            "type": "array",
            "minItems": 1,
            "title": "Evidence"
          },
          "external_action": {
            "type": "string",
            "enum": [
              "none",
              "calendar_draft",
              "checklist_update"
            ],
            "title": "External Action",
            "default": "none"
          },
          "requires_confirmation": {
            "type": "boolean",
            "title": "Requires Confirmation",
            "default": true
          },
          "prompt_version": {
            "type": "string",
            "minLength": 1,
            "title": "Prompt Version"
          }
        },
        "type": "object",
        "required": [
          "action_id",
          "title",
          "reason",
          "duration_minutes",
          "evidence",
          "prompt_version"
        ],
        "title": "ActionProposal",
        "description": "An evidence-backed next action that remains a proposal until approved."
      },
      "AgentRunCompleted": {
        "properties": {
          "status": {
            "type": "string",
            "const": "completed",
            "title": "Status"
          },
          "proposal": {
            "$ref": "#/components/schemas/ActionProposal"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "proposal"
        ],
        "title": "AgentRunCompleted"
      },
      "AgentRunRequest": {
        "properties": {
          "event": {
            "$ref": "#/components/schemas/OrbitEvent"
          },
          "context": {
            "items": {
              "$ref": "#/components/schemas/EvidenceLink"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Context"
          },
          "client_tools": {
            "items": {
              "$ref": "#/components/schemas/ClientTool"
            },
            "type": "array",
            "maxItems": 2,
            "title": "Client Tools"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "event",
          "context"
        ],
        "title": "AgentRunRequest",
        "description": "Start one explicit proposal run.\n\n``client_tools`` is only a capability advertisement.  The server never\nuses it to access Google; a connected extension must complete the deferred\ntool call and send the minimized derived result back."
      },
      "AgentRunToolRequired": {
        "properties": {
          "status": {
            "type": "string",
            "const": "tool_required",
            "title": "Status"
          },
          "run_id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Run Id"
          },
          "calls": {
            "items": {
              "$ref": "#/components/schemas/AgentToolCall"
            },
            "type": "array",
            "maxItems": 1,
            "minItems": 1,
            "title": "Calls"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "run_id",
          "calls"
        ],
        "title": "AgentRunToolRequired"
      },
      "AgentToolCall": {
        "properties": {
          "tool_call_id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Tool Call Id"
          },
          "name": {
            "type": "string",
            "enum": [
              "scombz_page_summary",
              "google_calendar_availability"
            ],
            "title": "Name"
          },
          "version": {
            "type": "integer",
            "const": 1,
            "title": "Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tool_call_id",
          "name",
          "version"
        ],
        "title": "AgentToolCall"
      },
      "AgentToolResultRequest": {
        "properties": {
          "tool_call_id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Tool Call Id"
          },
          "name": {
            "type": "string",
            "enum": [
              "scombz_page_summary",
              "google_calendar_availability"
            ],
            "title": "Name",
            "default": "google_calendar_availability"
          },
          "version": {
            "type": "integer",
            "const": 1,
            "title": "Version",
            "default": 1
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CalendarAvailabilityResult"
              },
              {
                "$ref": "#/components/schemas/ScombzPageSummaryResult"
              }
            ],
            "title": "Result"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tool_call_id",
          "result"
        ],
        "title": "AgentToolResultRequest",
        "description": "Result for one registered external tool.\n\nThe envelope repeats the call name and version so a result cannot be\naccidentally delivered to a different deferred tool.  The validator also\nkeeps the two strict result schemas from being mixed across tools."
      },
...
      "ChatHistoryMessage": {
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role"
          },
          "content": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Content"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "ChatHistoryMessage"
      },
      "ChatRunCompleted": {
        "properties": {
          "status": {
            "type": "string",
            "const": "completed",
            "title": "Status"
          },
          "message": {
            "$ref": "#/components/schemas/ChatAssistantMessage"
          },
          "proposal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActionProposal"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "title": "ChatRunCompleted"
      },
      "ChatRunRequest": {
        "properties": {
          "conversation_id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Conversation Id"
          },
          "message": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Message"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/ChatHistoryMessage"
            },
            "type": "array",
            "maxItems": 20,
            "title": "History"
          },
          "client_tools": {
            "items": {
              "$ref": "#/components/schemas/ChatClientTool"
            },
            "type": "array",
            "maxItems": 12,
            "title": "Client Tools"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "conversation_id",
          "message"
        ],
        "title": "ChatRunRequest"
      },
      "ChatRunToolRequired": {
        "properties": {
          "status": {
            "type": "string",
            "const": "tool_required",
            "title": "Status"
          },
          "run_id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Run Id"
          },
          "calls": {
            "items": {
              "$ref": "#/components/schemas/ChatToolCall"
            },
            "type": "array",
            "maxItems": 1,
            "minItems": 1,
            "title": "Calls"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "run_id",
          "calls"
        ],
        "title": "ChatRunToolRequired"
      },
...
      "ChatToolResultRequest": {
        "properties": {
          "tool_call_id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Tool Call Id"
          },
          "name": {
            "type": "string",
            "enum": [
              "scombz_page_summary",
              "scombz_read",
              "google_calendar_availability",
              "syllabus_search",
              "browser_read_url",
              "sitrus_read",
              "moodle_read",
              "my_library_read",
              "cast_read",
              "library_catalog_search",
              "library_item_read",
              "library_catalog_browse",
              "library_discovery_search"
            ],
            "title": "Name"
          },
          "version": {
            "type": "integer",
            "const": 1,
            "title": "Version"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CalendarAvailabilityResult"
              },
              {
                "$ref": "#/components/schemas/ScombzPageSummaryResult"
              },
              {
                "$ref": "#/components/schemas/ScombzReadResult"
              },
              {
                "$ref": "#/components/schemas/SyllabusSearchResult"
              },
              {
                "$ref": "#/components/schemas/BrowserReadResult"
              },
              {
                "$ref": "#/components/schemas/SitrusGradeResult"
              },
              {
                "$ref": "#/components/schemas/MoodleReadResult"
              },
              {
                "$ref": "#/components/schemas/MyLibraryReadResult"
              },
              {
                "$ref": "#/components/schemas/CastReadResult"
              },
              {
                "$ref": "#/components/schemas/LibraryCatalogSearchResult"
              },
              {
                "$ref": "#/components/schemas/LibraryItemReadResult"
              },
              {
                "$ref": "#/components/schemas/LibraryCatalogBrowseResult"
              },
              {
                "$ref": "#/components/schemas/LibraryDiscoverySearchResult"
              }
            ],
            "title": "Result"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tool_call_id",
          "name",
          "version",
          "result"
        ],
        "title": "ChatToolResultRequest"
      },
      "ClientTool": {
        "properties": {
          "name": {
            "type": "string",
            "enum": [
              "scombz_page_summary",
              "google_calendar_availability"
            ],
            "title": "Name"
          },
          "version": {
            "type": "integer",
            "const": 1,
            "title": "Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "version"
        ],
        "title": "ClientTool",
        "description": "A capability explicitly advertised by the client for one run."
      },
      "EvidenceLink": {
        "properties": {
          "evidence_id": {
            "type": "string",
            "minLength": 1,
            "title": "Evidence Id"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "title": "Title"
          },
          "source_type": {
            "type": "string",
            "enum": [
              "syllabus",
              "assignment",
              "learning_history",
              "calendar",
              "scombz",
              "library",
              "career",
              "google_drive",
              "web"
            ],
            "title": "Source Type"
          },
          "locator": {
            "type": "string",
            "minLength": 1,
            "title": "Locator"
          },
          "data_classification": {
            "type": "string",
            "enum": [
              "synthetic",
              "public",
              "personal",
              "restricted"
            ],
            "title": "Data Classification",
            "default": "synthetic"
          }
        },
        "type": "object",
        "required": [
          "evidence_id",
          "title",
          "source_type",
          "locator"
        ],
        "title": "EvidenceLink",
        "description": "A compact link to the evidence used by an action proposal."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrbitEvent": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "event_type": {
            "type": "string",
            "enum": [
              "campus_entered",
              "action_completed"
            ],
            "title": "Event Type"
          },
          "scenario_id": {
            "type": "string",
            "minLength": 1,
            "title": "Scenario Id"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "campus": {
            "type": "string",
            "enum": [
              "omiya",
              "toyosu",
              "other"
            ],
            "title": "Campus"
          },
          "data_classification": {
            "type": "string",
            "enum": [
              "synthetic",
              "public",
              "personal",
              "restricted"
            ],
            "title": "Data Classification",
            "default": "synthetic"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "event_type",
          "scenario_id",
          "campus"
        ],
        "title": "OrbitEvent",
        "description": "An observed event in the student's campus journey."
      },
      "ProposeActionRequest": {
        "properties": {
          "event": {
            "$ref": "#/components/schemas/OrbitEvent"
          },
          "context": {
            "items": {
              "$ref": "#/components/schemas/EvidenceLink"
            },
            "type": "array",
            "minItems": 1,
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "event",
          "context"
        ],
        "title": "ProposeActionRequest"
      },
      "ScombzPageSummaryResult": {
        "properties": {
          "route": {
            "type": "string",
            "enum": [
              "home",
              "tasks",
              "timetable",
              "announcements",
              "calendar",
              "course",
              "other"
            ],
            "title": "Route"
          },
          "task_count": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Task Count"
          },
          "announcement_count": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Announcement Count"
          },
          "related_link_count": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Related Link Count"
          },
          "has_current_course": {
            "type": "boolean",
            "title": "Has Current Course"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "route",
          "task_count",
          "announcement_count",
          "related_link_count",
          "has_current_course"
        ],
        "title": "ScombzPageSummaryResult",
        "description": "A minimized summary of the currently displayed ScombZ page.\n\nThis deliberately has no title, URL, course name, item, link, HTML, or\nbrowser token field.  The tool name and version are carried by the\nsurrounding deferred-tool result envelope."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerifyActionRequest": {
        "properties": {
          "scenario_id": {
            "type": "string",
            "minLength": 1,
            "title": "Scenario Id"
          },
          "campus": {
            "type": "string",
            "enum": [
              "omiya",
              "toyosu",
              "other"
            ],
            "title": "Campus"
          },
          "approved": {
            "type": "boolean",
            "title": "Approved"
          },
          "completed": {
            "type": "boolean",
            "title": "Completed"
          },
          "notes": {
            "type": "string",
            "maxLength": 500,
            "title": "Notes",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "scenario_id",
          "campus",
          "approved",
          "completed"
        ],
        "title": "VerifyActionRequest"
      }
    }
  }
}